Three function for follow
These codes realized the function for users to follow, unfollow, and store the updates from users' follows to further generate the notifications.

The queries are tested on Ratel and should work. But I am still writing the unit test to test them in the system. Would you please provide me with the "test_config.json" file for me to work on it?

Thanks!